tests/wpt/web-platform-tests/web-animations/animation/playbackRate.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Animation.playbackRate</title>
<link rel="help" href="https://w3c.github.io/web-animations/#dom-animation-playbackrate">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
"use strict";

function assert_playbackrate(animation,
                             previousAnimationCurrentTime,
                             previousTimelineCurrentTime,
                             description) {
  var accuracy = 0.001; /* accuracy of DOMHighResTimeStamp */
  var animationCurrentTimeDifference =
    animation.currentTime - previousAnimationCurrentTime;
  var timelineCurrentTimeDifference =
    animation.timeline.currentTime - previousTimelineCurrentTime;

  assert_approx_equals(animationCurrentTimeDifference,
                       timelineCurrentTimeDifference * animation.playbackRate,
                       accuracy,
                       description);
}

promise_test(function(t) {
  var div = createDiv(t);
  var animation = div.animate(null, 100 * MS_PER_SEC);
  return animation.ready.then(function() {
    animation.currentTime = 7 * MS_PER_SEC; // ms
    animation.playbackRate = 0.5;

    assert_equals(animation.currentTime, 7 * MS_PER_SEC,
      'Reducing Animation.playbackRate should not change the currentTime ' +
      'of a playing animation');
    animation.playbackRate = 2;
    assert_equals(animation.currentTime, 7 * MS_PER_SEC,
      'Increasing Animation.playbackRate should not change the currentTime ' +
      'of a playing animation');
  });
}, 'Test the initial effect of setting playbackRate on currentTime');

promise_test(function(t) {
  var div = createDiv(t);
  var animation = div.animate(null, 100 * MS_PER_SEC);
  animation.playbackRate = 2;
  var previousTimelineCurrentTime;
  var previousAnimationCurrentTime;
  return animation.ready.then(function() {
    previousAnimationCurrentTime = animation.currentTime;
    previousTimelineCurrentTime = animation.timeline.currentTime;
    return waitForAnimationFrames(1);
  }).then(function() {
    assert_playbackrate(animation,
      previousAnimationCurrentTime,
      previousTimelineCurrentTime,
      'animation.currentTime should be 2 times faster than timeline.');
  });
}, 'Test the effect of setting playbackRate on currentTime');

promise_test(function(t) {
  var div = createDiv(t);
  var animation = div.animate(null, 100 * MS_PER_SEC);
  animation.playbackRate = 2;
  var previousTimelineCurrentTime;
  var previousAnimationCurrentTime;
  return animation.ready.then(function() {
    previousAnimationCurrentTime = animation.currentTime;
    previousTimelineCurrentTime = animation.timeline.currentTime;
    animation.playbackRate = 1;
    return waitForAnimationFrames(1);
  }).then(function() {
    assert_equals(animation.playbackRate, 1,
      'sanity check: animation.playbackRate is still 1.');
    assert_playbackrate(animation,
      previousAnimationCurrentTime,
      previousTimelineCurrentTime,
      'animation.currentTime should be the same speed as timeline now.');
  });
}, 'Test the effect of setting playbackRate while playing animation');

</script>
</body>
